chore: add Huri model
